+++ pages/candidate/[displayName]/edit-profile.js
@@ -1,43 +1,33 @@
+// React/Next imports
 import { useState, useEffect } from 'react'
-import { useRouter } from 'next/router'
-import { db } from 'utils/db'
 import { useAuth } from 'store/AuthContext'
-// import { useProfileInfo } from 'store/profile_store'
+import { useProfileInfo } from 'store/profile_info'
+import { useRouter } from 'next/router'
+import NavLink from 'components/global/NavLink'
+
+// Lib Imports
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
+import { db } from 'utils/db'
 import * as yup from 'yup'
 import getText from 'utils/i18n/Texts'
 import ProfileMenu from 'components/user/ProfileMenu'
-import timezones from 'data/timezones.json'
-import { getStaticProps } from 'pages/learning-resources'
 
-const CandidateEditProfile = ({ profileInfo }) => {
+import timezones from 'data/timezones.json'
+
+const CandidateEditProfile = () => {
   const router = useRouter()
   const { currentUser } = useAuth()
-  // const [profileInfo, setProfileInfo] = useState({})
-  // const profile = useProfileInfo((s) => s.profile)
   const [loading, setLoading] = useState(true)
   const [error, setError] = useState(null)
   const [timezonesArray, setTimezonesArray] = useState([])
+  const profileInfo = useProfileInfo((s) => s.profileInfo)
 
-  console.log('Profile Info:', profileInfo)
+  const displayNameUrl = router.query.displayName
 
   useEffect(() => {
     setTimezonesArray(timezones)
   }, [])
-
-  // useEffect(() => {
-  //   db.collection('candidates')
-  //     .where('uid', '==', currentUser.uid)
-  //     .get()
-
-  //     .then((querySnapshot) => {
-  //       if (!querySnapshot.empty) {
-  //         console.log('Profile Info:', querySnapshot.docs[0].data())
-  //         setProfileInfo(querySnapshot.docs[0].data())
-  //       }
-  //     })
-  // }, [])
 
   const Schema = yup.object().shape({
     firstName: yup.string().required(getText('ACCOUNT', 'FIRST_NAME_REQUIRED')),
@@ -82,19 +72,18 @@ const CandidateEditProfile = ({ profileInfo }) => {
 
   const handleProfileForm = (data) => {
     setLoading(true)
-    db.collection('candidates')
-      .add({
-        uid: currentUser.uid,
+    db.collection('candidate')
+      .doc(currentUser.uid)
+      .update({
+        accountType: 'candidate',
         firstName: data.firstName,
         lastName: data.lastName,
         email: data.email,
         portfolio: data.portfolio,
-        social: {
-          dev: data.social_dev,
-          github: data.social_github,
-          linkedin: data.social_linkedin,
-          twitter: data.social_twitter,
-        },
+        dev: data.social_dev,
+        github: data.social_github,
+        linkedin: data.social_linkedin,
+        twitter: data.social_twitter,
         hideInfo: data.hideInfo,
         timezone: data.timezone,
         timeframe_from: data.timeframe_from,
@@ -114,351 +103,348 @@ const CandidateEditProfile = ({ profileInfo }) => {
 
   return (
     <>
-      {currentUser ? (
-        <div className='max-w-4xl mx-auto'>
-          <div className='flex flex-col justify-between space-y-6 md:space-x-24 md:space-y-0 md:flex-row'>
-            <div className='flex flex-col order-2 space-y-4 md:order-1'>
-              <ProfileMenu />
+      <div className='max-w-4xl mx-auto'>
+        <div className='flex flex-col justify-between space-y-6 md:space-x-24 md:space-y-0 md:flex-row'>
+          <div className='flex flex-col order-2 space-y-4 md:order-1'>
+            <ProfileMenu>
+              <li className='font-bold'>{currentUser.email}</li>
+              <li>
+                <NavLink
+                  href={`/candidate/${displayNameUrl}/`}
+                  activeClassName='text-teal-500'
+                >
+                  {getText('ACCOUNT', 'VIEW_PROFILE')}
+                </NavLink>
+              </li>
+              <li>
+                <NavLink
+                  href={`/candidate/${displayNameUrl}/edit-profile`}
+                  activeClassName='text-teal-500 font-bold'
+                >
+                  {getText('ACCOUNT', 'EDIT_PROFILE')}
+                </NavLink>
+              </li>
+            </ProfileMenu>
+          </div>
+
+          <div className='order-1 w-full md:order-2'>
+            <div className='flex items-center justify-between mb-6'>
+              <h1 className='mb-3 text-lg text-blue-900'>
+                {getText('ACCOUNT', 'PROFILE_INFO')}
+              </h1>
             </div>
 
-            <div className='order-1 w-full md:order-2'>
-              <div className='flex items-center justify-between mb-6'>
-                <h1 className='mb-3 text-lg text-blue-900'>
-                  {getText('ACCOUNT', 'PROFILE_INFO')}
-                </h1>
+            <form
+              autoComplete='on'
+              onSubmit={handleSubmit(handleProfileForm)}
+              className='mb-6'
+            >
+              {/* name */}
+              <div className='flex flex-col justify-between md:space-x-8 md:flex-row'>
+                <div className='flex flex-col w-full mb-3 space-y-1'>
+                  <label htmlFor='firstName'>
+                    {getText('ACCOUNT', 'FIRST_NAME')}
+                  </label>
+                  <input
+                    id='firstName'
+                    type='text'
+                    name='firstName'
+                    className='input'
+                    ref={register}
+                  />
+                  {errors.firstName ? (
+                    <p className='input-error'>
+                      {errors.firstName && errors.firstName.message}
+                    </p>
+                  ) : null}
+                </div>
+
+                <div className='flex flex-col w-full mb-3 space-y-1'>
+                  <label htmlFor='lastName'>
+                    {getText('ACCOUNT', 'LAST_NAME')}
+                  </label>
+                  <input
+                    id='lastName'
+                    type='text'
+                    name='lastName'
+                    className='input'
+                    ref={register}
+                  />
+                  {errors.lastName ? (
+                    <p className='input-error'>
+                      {errors.lastName && errors.lastName.message}
+                    </p>
+                  ) : null}
+                </div>
               </div>
 
-              <form
-                autoComplete='on'
-                onSubmit={handleSubmit(handleProfileForm)}
-                className='mb-6'
-              >
-                {/* name */}
-                <div className='flex flex-col justify-between md:space-x-8 md:flex-row'>
-                  <div className='flex flex-col w-full mb-3 space-y-1'>
-                    <label htmlFor='firstName'>
-                      {getText('ACCOUNT', 'FIRST_NAME')}
-                    </label>
+              {/* email/portfolio */}
+              <div className='flex flex-col justify-between mb-8 md:space-x-8 md:flex-row'>
+                <div className='flex flex-col w-full mb-3 space-y-1'>
+                  <label htmlFor='email'>{getText('ACCOUNT', 'EMAIL')}</label>
+                  <input
+                    id='email'
+                    type='text'
+                    name='email'
+                    className='input'
+                    ref={register}
+                  />
+                  {errors.email ? (
+                    <p className='input-error'>
+                      {errors.email && errors.email.message}
+                    </p>
+                  ) : null}
+                </div>
+
+                <div className='flex flex-col w-full mb-3 space-y-1'>
+                  <label htmlFor='portfolio'>
+                    {getText('ACCOUNT', 'PORTFOLIO')}
+                    <span className='text-sm font-normal'> (optional)</span>
+                  </label>
+                  <input
+                    id='portfolio'
+                    type='text'
+                    name='portfolio'
+                    className='input'
+                    ref={register}
+                  />
+                </div>
+              </div>
+
+              {/* social/timezone */}
+              <div className='flex flex-col justify-between mb-12 space-y-6 md:space-y-0 md:space-x-8 md:flex-row'>
+                <div className='flex flex-col w-full space-y-4'>
+                  <label htmlFor='social' className='mb-0'>
+                    {getText('ACCOUNT', 'SOCIAL_ACCOUNTS')}
+                  </label>
+                  <div className='flex items-center space-x-3'>
+                    <div className='text-3xl opacity-50'>
+                      <i className='fab fa-dev'></i>
+                    </div>
                     <input
-                      id='firstName'
                       type='text'
-                      name='firstName'
-                      className='input'
+                      name='social_dev'
+                      className='w-full input'
                       ref={register}
                     />
-                    {errors.firstName ? (
-                      <p className='input-error'>
-                        {errors.firstName && errors.firstName.message}
-                      </p>
-                    ) : null}
                   </div>
 
-                  <div className='flex flex-col w-full mb-3 space-y-1'>
-                    <label htmlFor='lastName'>
-                      {getText('ACCOUNT', 'LAST_NAME')}
-                    </label>
+                  <div className='flex items-center space-x-2'>
+                    <div className='text-3xl opacity-50'>
+                      <i className='fab fa-github'></i>
+                    </div>
                     <input
-                      id='lastName'
                       type='text'
-                      name='lastName'
-                      className='input'
+                      name='social_github'
+                      className='w-full input '
                       ref={register}
                     />
-                    {errors.lastName ? (
-                      <p className='input-error'>
-                        {errors.lastName && errors.lastName.message}
-                      </p>
-                    ) : null}
+                  </div>
+
+                  <div className='flex items-center space-x-3'>
+                    <div className='text-3xl opacity-50'>
+                      <i className='fab fa-linkedin'></i>
+                    </div>
+                    <input
+                      type='text'
+                      name='social_linkedin'
+                      className='w-full input '
+                      ref={register}
+                    />
+                  </div>
+
+                  <div className='flex items-center space-x-2'>
+                    <div className='text-3xl opacity-50'>
+                      <i className='fab fa-twitter'></i>
+                    </div>
+                    <input
+                      type='text'
+                      name='social_twitter'
+                      className='w-full input '
+                      ref={register}
+                    />
                   </div>
                 </div>
 
-                {/* email/portfolio */}
-                <div className='flex flex-col justify-between mb-8 md:space-x-8 md:flex-row'>
-                  <div className='flex flex-col w-full mb-3 space-y-1'>
-                    <label htmlFor='email'>{getText('ACCOUNT', 'EMAIL')}</label>
-                    <input
-                      id='email'
+                <div className='flex flex-col justify-between w-full space-y-4'>
+                  {/* hide info */}
+                  <div className='flex items-center justify-between mb-3'>
+                    <label htmlFor='hideInfo' className='mb-0'>
+                      {getText('ACCOUNT', 'HIDE_INFO')}
+                    </label>
+                    <div className='relative inline-block w-12 align-middle transition duration-700 ease-in-out select-none'>
+                      <input
+                        type='checkbox'
+                        name='hideInfo'
+                        id='toggle'
+                        className='absolute block w-5 h-5 bg-white border-gray-300 rounded-full outline-none appearance-none cursor-pointer border-3 toggle-checkbox'
+                        ref={register}
+                      />
+                      <label
+                        htmlFor='toggle'
+                        className='block h-5 mb-0 overflow-hidden bg-gray-300 rounded-full cursor-pointer toggle-label'
+                      ></label>
+                    </div>
+                  </div>
+
+                  {/* timezone */}
+                  <div className='flex flex-col w-full space-y-1'>
+                    <label htmlFor='timezone'>
+                      {getText('ACCOUNT', 'TIMEZONE')}
+                    </label>
+                    <select
+                      id='timezone'
                       type='text'
-                      name='email'
-                      className='input'
+                      name='timezone'
+                      className='w-full input'
                       ref={register}
-                    />
-                    {errors.email ? (
+                    >
+                      <option value={getText('DASHBOARD', 'SELECT')}>
+                        {getText('DASHBOARD', 'SELECT')}
+                      </option>
+                      {timezonesArray.map((timezone, index) => {
+                        return (
+                          <option
+                            key={index}
+                            value={timezone.text}
+                            className='text-gray-300'
+                          >
+                            {timezone.text}
+                          </option>
+                        )
+                      })}
+                    </select>
+                    {errors.timezone ? (
                       <p className='input-error'>
-                        {errors.email && errors.email.message}
+                        {errors.timezone && errors.timezone.message}
                       </p>
                     ) : null}
                   </div>
 
-                  <div className='flex flex-col w-full mb-3 space-y-1'>
-                    <label htmlFor='portfolio'>
-                      {getText('ACCOUNT', 'PORTFOLIO')}
-                      <span className='text-sm font-normal'> (optional)</span>
-                    </label>
-                    <input
-                      id='portfolio'
-                      type='text'
-                      name='portfolio'
-                      className='input'
-                      ref={register}
-                    />
-                  </div>
-                </div>
-
-                {/* social/timezone */}
-                <div className='flex flex-col justify-between mb-12 space-y-6 md:space-y-0 md:space-x-8 md:flex-row'>
+                  {/* timeframe */}
                   <div className='flex flex-col w-full space-y-4'>
-                    <label htmlFor='social' className='mb-0'>
-                      {getText('ACCOUNT', 'SOCIAL_ACCOUNTS')}
+                    <label htmlFor='timeframe'>
+                      {getText('ACCOUNT', 'TIMEFRAME')}
                     </label>
-                    <div className='flex items-center space-x-3'>
-                      <div className='text-3xl opacity-50'>
-                        <i className='fab fa-dev'></i>
-                      </div>
-                      <input
-                        type='text'
-                        name='social_dev'
-                        className='w-full input'
-                        ref={register}
-                      />
-                    </div>
 
-                    <div className='flex items-center space-x-2'>
-                      <div className='text-3xl opacity-50'>
-                        <i className='fab fa-github'></i>
-                      </div>
-                      <input
-                        type='text'
-                        name='social_github'
-                        className='w-full input '
-                        ref={register}
-                      />
-                    </div>
-
-                    <div className='flex items-center space-x-3'>
-                      <div className='text-3xl opacity-50'>
-                        <i className='fab fa-linkedin'></i>
-                      </div>
-                      <input
-                        type='text'
-                        name='social_linkedin'
-                        className='w-full input '
-                        ref={register}
-                      />
-                    </div>
-
-                    <div className='flex items-center space-x-2'>
-                      <div className='text-3xl opacity-50'>
-                        <i className='fab fa-twitter'></i>
-                      </div>
-                      <input
-                        type='text'
-                        name='social_twitter'
-                        className='w-full input '
-                        ref={register}
-                      />
-                    </div>
-                  </div>
-
-                  <div className='flex flex-col justify-between w-full space-y-4'>
-                    {/* hide info */}
-                    <div className='flex items-center justify-between mb-3'>
-                      <label htmlFor='hideInfo' className='mb-0'>
-                        {getText('ACCOUNT', 'HIDE_INFO')}
-                      </label>
-                      <div className='relative inline-block w-12 align-middle transition duration-700 ease-in-out select-none'>
-                        <input
-                          type='checkbox'
-                          name='hideInfo'
-                          id='toggle'
-                          className='absolute block w-5 h-5 bg-white border-gray-300 rounded-full outline-none appearance-none cursor-pointer border-3 toggle-checkbox'
+                    <div className='flex justify-between space-x-3'>
+                      <div className='flex flex-col w-full space-y-1'>
+                        <select
+                          id='timeframe_from'
+                          type='text'
+                          name='timeframe_from'
+                          className='input'
                           ref={register}
-                        />
-                        <label
-                          htmlFor='toggle'
-                          className='block h-5 mb-0 overflow-hidden bg-gray-300 rounded-full cursor-pointer toggle-label'
-                        ></label>
+                        >
+                          <option value=''>From</option>
+                          <option value='Option #1'>Option #1</option>
+                        </select>
+                        {errors.timeframe_from ? (
+                          <p className='input-error'>
+                            {errors.timeframe_from &&
+                              errors.timeframe_from.message}
+                          </p>
+                        ) : null}
                       </div>
-                    </div>
 
-                    {/* timezone */}
-                    <div className='flex flex-col w-full space-y-1'>
-                      <label htmlFor='timezone'>
-                        {getText('ACCOUNT', 'TIMEZONE')}
-                      </label>
-                      <select
-                        id='timezone'
-                        type='text'
-                        name='timezone'
-                        className='w-full input'
-                        ref={register}
-                      >
-                        <option value={getText('DASHBOARD', 'SELECT')}>
-                          {getText('DASHBOARD', 'SELECT')}
-                        </option>
-                        {timezonesArray.map((timezone, index) => {
-                          return (
-                            <option
-                              key={index}
-                              value={timezone.text}
-                              className='text-gray-300'
-                            >
-                              {timezone.text}
-                            </option>
-                          )
-                        })}
-                      </select>
-                      {errors.timezone ? (
-                        <p className='input-error'>
-                          {errors.timezone && errors.timezone.message}
-                        </p>
-                      ) : null}
-                    </div>
+                      <div className='flex flex-col w-full space-y-1'>
+                        <select
+                          id='timeframe_to'
+                          type='text'
+                          name='timeframe_to'
+                          className='input'
+                          ref={register}
+                        >
+                          <option value=''>To</option>
+                          <option value='Option #1'>Option #1</option>
+                        </select>
 
-                    {/* timeframe */}
-                    <div className='flex flex-col w-full space-y-4'>
-                      <label htmlFor='timeframe'>
-                        {getText('ACCOUNT', 'TIMEFRAME')}
-                      </label>
-
-                      <div className='flex justify-between space-x-3'>
-                        <div className='flex flex-col w-full space-y-1'>
-                          <select
-                            id='timeframe_from'
-                            type='text'
-                            name='timeframe_from'
-                            className='input'
-                            ref={register}
-                          >
-                            <option value=''>From</option>
-                            <option value='Option #1'>Option #1</option>
-                          </select>
-                          {errors.timeframe_from ? (
-                            <p className='input-error'>
-                              {errors.timeframe_from &&
-                                errors.timeframe_from.message}
-                            </p>
-                          ) : null}
-                        </div>
-
-                        <div className='flex flex-col w-full space-y-1'>
-                          <select
-                            id='timeframe_to'
-                            type='text'
-                            name='timeframe_to'
-                            className='input'
-                            ref={register}
-                          >
-                            <option value=''>To</option>
-                            <option value='Option #1'>Option #1</option>
-                          </select>
-
-                          {errors ? (
-                            <p className='input-error'>
-                              {errors.timeframe_to &&
-                                errors.timeframe_to.message}
-                            </p>
-                          ) : null}
-                        </div>
+                        {errors ? (
+                          <p className='input-error'>
+                            {errors.timeframe_to && errors.timeframe_to.message}
+                          </p>
+                        ) : null}
                       </div>
                     </div>
                   </div>
                 </div>
+              </div>
 
-                {/* question1 */}
-                <div className='flex flex-col w-full mb-10 space-y-3'>
-                  <label htmlFor='question1'>
-                    {getText('ACCOUNT', 'QUESTION1')}
-                  </label>
-                  <textarea
-                    name='question1'
-                    id='question1'
-                    className='input'
-                    cols='30'
-                    rows='5'
-                    ref={register}
-                  ></textarea>
-                  {errors ? (
-                    <p className='input-error'>
-                      {errors.question1 && errors.question1.message}
-                    </p>
-                  ) : null}
-                </div>
-
-                {/* question2 */}
-                <div className='flex flex-col w-full mb-10 space-y-3'>
-                  <label htmlFor='question2'>
-                    {getText('ACCOUNT', 'QUESTION2')}
-                  </label>
-                  <textarea
-                    name='question2'
-                    id='question2'
-                    className='input'
-                    cols='30'
-                    rows='5'
-                    ref={register}
-                  ></textarea>
-                  {errors ? (
-                    <p className='input-error'>
-                      {errors.question2 && errors.question2.message}
-                    </p>
-                  ) : null}
-                </div>
-
-                {/* question3 */}
-                <div className='flex flex-col w-full mb-4 space-y-3'>
-                  <label htmlFor='question3'>
-                    {getText('ACCOUNT', 'QUESTION3')}
-                  </label>
-                  <textarea
-                    name='question3'
-                    id='question3'
-                    className='input'
-                    cols='30'
-                    rows='5'
-                    ref={register}
-                  ></textarea>
-                  {errors ? (
-                    <p className='input-error'>
-                      {errors.question3 && errors.question3.message}
-                    </p>
-                  ) : null}
-                </div>
-
-                <button type='submit' className='btn btn-teal'>
-                  {getText('ACCOUNT', 'SAVE')}
-                </button>
-
-                {error ? (
-                  <p className='p-3 mt-6 text-lg text-center text-red-500 bg-red-100 rounded-md'>
-                    {error}
+              {/* question1 */}
+              <div className='flex flex-col w-full mb-10 space-y-3'>
+                <label htmlFor='question1'>
+                  {getText('ACCOUNT', 'QUESTION1')}
+                </label>
+                <textarea
+                  name='question1'
+                  id='question1'
+                  className='input'
+                  cols='30'
+                  rows='5'
+                  ref={register}
+                ></textarea>
+                {errors ? (
+                  <p className='input-error'>
+                    {errors.question1 && errors.question1.message}
                   </p>
                 ) : null}
-              </form>
-            </div>
+              </div>
+
+              {/* question2 */}
+              <div className='flex flex-col w-full mb-10 space-y-3'>
+                <label htmlFor='question2'>
+                  {getText('ACCOUNT', 'QUESTION2')}
+                </label>
+                <textarea
+                  name='question2'
+                  id='question2'
+                  className='input'
+                  cols='30'
+                  rows='5'
+                  ref={register}
+                ></textarea>
+                {errors ? (
+                  <p className='input-error'>
+                    {errors.question2 && errors.question2.message}
+                  </p>
+                ) : null}
+              </div>
+
+              {/* question3 */}
+              <div className='flex flex-col w-full mb-4 space-y-3'>
+                <label htmlFor='question3'>
+                  {getText('ACCOUNT', 'QUESTION3')}
+                </label>
+                <textarea
+                  name='question3'
+                  id='question3'
+                  className='input'
+                  cols='30'
+                  rows='5'
+                  ref={register}
+                ></textarea>
+                {errors ? (
+                  <p className='input-error'>
+                    {errors.question3 && errors.question3.message}
+                  </p>
+                ) : null}
+              </div>
+
+              <button type='submit' className='btn btn-teal'>
+                {getText('ACCOUNT', 'SAVE')}
+              </button>
+
+              {error ? (
+                <p className='p-3 mt-6 text-lg text-center text-red-500 bg-red-100 rounded-md'>
+                  {error}
+                </p>
+              ) : null}
+            </form>
           </div>
         </div>
-      ) : null}
+      </div>
     </>
   )
-}
-
-export async function getServerSideProps(context) {
-  let profileInfo
-  const querySnapshot = await db
-    .collection('candidates')
-    .where('uid', '==', 'Rd6ApvzQ0I2ViWqdfUM0CU1mvRMu')
-    .get()
-
-  if (!querySnapshot.empty) {
-    // console.log('Profile Info:', querySnapshot.docs[0].data())
-    profileInfo = await querySnapshot.docs[0].data()
-  }
-  return {
-    props: {
-      profileInfo,
-    },
-  }
 }
 
 export default CandidateEditProfile
